requirements.txt
async-timeout==3.0.1
asyncio==3.4.3
certifi==2021.5.30
configparser==5.0.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.3
httpx==0.18.1
httpx-socks==0.4.0
hyperframe==5.2.0
idna==3.2
pyaes==1.6.1
pyasn1==0.4.8
python-crontab==2.5.1
python-dateutil==2.8.1
python-socks==1.2.4
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
Telethon==1.21.1
